+++ optimum/neuron/pipelines/diffusers/__init__.py
@@ -12,3 +12,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from .pipeline_utils import StableDiffusionXLPipelineMixin
